# contracts/openapi/employer-service.yaml
...
openapi: 3.1.0
info:
  title: employer-service API
  version: 0.1.0-gate0
  description: 'Generated by docs/tools/build_gate0.py — do not edit by hand; edit the sources and re-run.


    SYNTHETIC DEMONSTRATION — NOT AN OFFICIAL EPFO SYSTEM. x-status: working | mock | planned | definition-pending. Planned
    operations return 501 /problems/planned.'
servers:
- url: http://localhost:8000/api/v1
  description: Gateway (init.md §2.4)
paths:
...
  /integrations/mca/registrations:
    post:
      operationId: postIntegrationsMcaRegistrations
      summary: MCA SPICe+ / AGILE-PRO auto-registration feed (signed)
      tags:
      - Establishment registration and configuration
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-status: mock
      x-phase: 1
      x-owner-service: employer
      x-callers:
      - ext.mca
      x-scope: signed partner callback (signature, event ID, replay window)
      security:
      - clientCredentials: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: Defined when implemented
  /integrations/shram-suvidha/registrations:
    post:
      operationId: postIntegrationsShramSuvidhaRegistrations
      summary: Shram Suvidha common-registration feed (signed)
      tags:
      - Establishment registration and configuration
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-status: mock
      x-phase: 1
      x-owner-service: employer
      x-callers:
      - ext.shram_suvidha
      x-scope: signed partner callback (signature, event ID, replay window)
      security:
      - clientCredentials: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: Defined when implemented
...
components:
  securitySchemes:
    bffSession:
      type: apiKey
      in: cookie
      name: __Host-epfo-session
      description: Gateway session cookie (init.md §2.1, §6.6)
    clientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8080/realms/epfo-demo/protocol/openid-connect/token
          scopes: {}
  parameters:
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required:
      - type
      - title
      - status
      - correlation_id
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        correlation_id:
          type: string
          format: uuid
    Meta:
      type: object
      required:
      - correlation_id
      - api_version
      - source
      - as_of
      properties:
        correlation_id:
          type: string
          format: uuid
        api_version:
          const: v1
        source:
          const: synthetic-poc
        as_of:
          type: string
          format: date-time
    Envelope:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: object
          description: Resource-specific; defined when the endpoint is implemented
        meta:
          $ref: '#/components/schemas/Meta'
  responses:
...
    Created:
      description: Created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...